# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  routes.CommentDTO:
    properties:
      authorId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      isReply:
        type: boolean
      postId:
        type: string
      replyToId:
        type: string
      reputation:
        type: integer
    type: object
...
  routes.CreatePostRequest:
    properties:
      content:
        type: string
      groupId:
        type: string
      mediaUrls:
        items:
          type: string
        type: array
    required:
    - content
    type: object
...
  routes.PaginatedPostsResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      posts:
        items:
          $ref: '#/definitions/routes.PostDTO'
        type: array
      totalCount:
        type: integer
    type: object
  routes.PostDTO:
    properties:
      authorId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      groupId:
        type: string
      id:
        type: string
      mediaUrls:
        items:
          type: string
        type: array
      reputation:
        type: integer
    type: object
  routes.PostDetailDTO:
    properties:
      authorId:
        type: string
      comments:
        items:
          $ref: '#/definitions/routes.CommentDTO'
        type: array
      content:
        type: string
      createdAt:
        type: string
      groupId:
        type: string
      id:
        type: string
      mediaUrls:
        items:
          type: string
        type: array
      reputation:
        type: integer
    type: object
...
  routes.UpdatePostRequest:
    properties:
      content:
        type: string
      mediaUrls:
        items:
          type: string
        type: array
    type: object
...
  routes.VoteRequest:
    properties:
      value:
        enum:
        - -1
        - 1
        type: integer
    required:
    - value
    type: object
  routes.VoteResponse:
    properties:
      reputation:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a Reddit-like REST API.
  title: Chirp API
  version: "1.0"
paths:
...
  /groups/{groupId}/subscribe:
    delete:
      description: Отписывает пользователя от группы
      parameters:
      - description: ID группы
        in: path
        name: groupId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отписаться от группы
      tags:
      - subscriptions
    post:
      description: Подписывает пользователя на группу
      parameters:
      - description: ID группы
        in: path
        name: groupId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Подписаться на группу
      tags:
      - subscriptions
...
  /posts:
    get:
      description: Получает посты с пагинацией
      parameters:
      - description: Страница
        in: query
        name: page
        type: integer
      - description: Лимит
        in: query
        name: limit
        type: integer
      - description: Сортировка (createdAt|reputation)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.PaginatedPostsResponse'
      summary: Получить список постов
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Создаёт новый пост
      parameters:
      - description: Данные для поста
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.PostDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать пост
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Удаляет пост пользователя
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить пост
      tags:
      - posts
    get:
      description: Получает подробную информацию о посте
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.PostDetailDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить детали поста
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Обновляет пост пользователя
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.PostDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить пост
      tags:
      - posts
  /posts/{id}/vote:
    post:
      consumes:
      - application/json
      description: Голосует за пост (лайк/дизлайк)
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      - description: Голос
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.VoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.VoteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Голосовать за пост
      tags:
      - posts
...
swagger: "2.0"